ISSUE #3682 - move reusable functions into helper
